🐛 Fix deployment by using assumed AWS role
In #485 we optimised the build by merging two jobs into one: assuming
the deployment role, and carrying out the deployment. The two steps
passed an AWS credentials file between them using workspace persistence.
Merging the jobs removed some spin-up time from the overall run.

However, in
https://app.circleci.com/pipelines/github/LBHackney-IT/lbh-housing-register/1613/workflows/68ba707c-0862-45cd-a55d-6940a82a64eb/jobs/7067/parallel-runs/0/steps/0-107
the deployment step failed because Serverless was unable to retrieve
secrets from AWS.

It turns out this failure cropped up because the
lbh-hackit/aws_assume_role orb uses environment variables but outputs
a credentials file. Environment variables take precedence, therefore the
orb effectively no-ops unless a clean environment is present.

Ideally we would/could/should fix this upstream. There are two promising
routes:

1. Update the LBH orb to use the new role credentials (i.e. unset the
   environment variables _or_ update them to match the assumed role
   creds); or

2. Use the aws-cli orb's setup and role assumption commands directly,
   and deprecate the LBH orb

Option 2 is (probably) the better one, as it removes
a Hackney-maintained dependency.

I've not attempted this as part of this change as the goal is to make
the build succeed again. I've also not refactored the three similar
deployment jobs into a common command to reduce duplication for the same
reason.