ci: Unset env var in CI build step
In #487 we unset the AWS credential environment variables before running
the deploy step, so that `aws-cli` would use the credentials stored on
disk by the assume-role step.

However, it looks like the environment variables may be reset for each
step, because although unsetting them in an SSH debug session succeeded,
the deployment failed:
https://app.circleci.com/pipelines/github/LBHackney-IT/lbh-housing-register/1616/workflows/6a32b504-27fa-4824-b702-59fce56128e0/jobs/7102.

If that assumption is correct, this change will resolve the deployment
failure by unsetting the environment variable in the same bash shell as
the deploy command is run.

If this change continues to fail, then #487 wasn't an effective fix and
we can revert both this and that change.